Create actions and reducers via createSlice
