+++ src/ProjectMain.tsx
@@ -14,7 +14,7 @@ import { getAppContextUtil } from './util/getAppContextUtil';
 import SiderMenu from './component/SiderMenu/SiderMenu';
 import Footer from './component/container/Footer/Footer';
 import AddLayerPanel from './component/AddLayerPanel/AddLayerPanel';
-import { hideAddLayerWindow } from './state/appState/actions';
+import { hideAddLayerWindow } from './state/appState';
 
 import PermalinkUtil from '@terrestris/ol-util/dist/PermalinkUtil/PermalinkUtil';
