--- src/state/appInfo/actions.ts
@@ -1,8 +1,0 @@
-import { createAction } from '@reduxjs/toolkit';
-
-export interface AppInfo {
-  name: string;
-  versionNumber: string;
-};
-
-export const addAppInfo = createAction<any>('appInfo/add');
