+++ src/state/appInfo/index.ts
@@ -1,0 +1,30 @@
+import {
+  createSlice,
+  PayloadAction
+} from '@reduxjs/toolkit';
+
+export interface AppInfo {
+  name: string;
+  versionNumber: string;
+};
+
+const initialState: AppInfo = {
+  name: '',
+  versionNumber: ''
+};
+
+const slice = createSlice({
+  name: 'appInfo',
+  initialState,
+  reducers: {
+    addAppInfo(state, action: PayloadAction<AppInfo>) {
+      state = action.payload;
+    }
+  }
+});
+
+export const {
+  addAppInfo
+} = slice.actions;
+
+export default slice.reducer;
